requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram-inline-paginations==0.1.12
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-translate==3.11.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
